+++ src/models/index.ts
@@ -5,3 +5,5 @@ export * from "./state.model.js";
 export * from "./stats.model.js";
 export * from "./strategy.model.js";
 export * from "./task.model.js";
+export * from "./events.model.js";
+